+++ tests/testthat/test-class-workbook-wrappers.R
@@ -265,12 +265,13 @@ test_that("wb_set_grid_lines() is a wrapper", {
 
 # wb_add_named_region(), wb_remove_named_region() -------------------------
 
-test_that("wb_add_named_region(), wb_remove_named_region() are wrappers", {
+test_that("wb_add_named_region(), wb_remove_named_region() wb_get_named_regions() are wrappers", {
   wb <- wb_workbook()$add_worksheet("a")
   params <- list(sheet = 1, dims = "A1", name = "cool")
   expect_wrapper("add_named_region", wb = wb, params = params)
   # now add the named region so that we can remove it
   wb$add_named_region(sheet = 1, dims = "A1", name = "cool")
+  expect_wrapper("get_named_regions", wb = wb)
   expect_wrapper("remove_named_region", wb = wb, params = list(name = "cool"))
 })
 
@@ -382,7 +383,16 @@ test_that("wb_add_comment() is a wrapper", {
     wb = wb,
     params = list(comment = c1, dims = "A1")
   )
+  opt <- getOption("openxlsx2.creator")
+  options("openxlsx2.creator" = "user")
+  wb <- wb_workbook()$add_worksheet()
 
+  expect_wrapper(
+    "add_comment",
+    wb = wb,
+    params = list(comment = "a new comment", dims = "A1")
+  )
+  options("openxlsx2.creator" = opt)
 })
 
 # wb_remove_comment() -----------------------------------------------------